integration/pyproject.toml
[tool.poetry]
name = "mip"
version = "0.1.0"
description = ""
authors = ["Michael P. Gerlek <mgerlek@inferlink.com>"]
readme = "README.md"

[tool.poetry.dependencies]
nvidia-ml-py3 = "7.352.0"
python = "^3.11"
psutil = "^5.9.7"
pydantic = "^2.5.2"
pydot = "^2.0.0"
pydantic-yaml = "^1.2.0"
filelock = "^3.13.1"
luigi = "^3.4.0"
pyyaml = "^6.0.1"
docker = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
